+++ src/clflags.ml
@@ -7,3 +7,5 @@ let warnings = ref "-40"
 let debug_dep_path = ref false
 let dev_mode = ref false
 let workspace_root = ref "."
+let external_lib_deps_hint = ref []
+